Upgrade to Quarkus 3.0
the main change is migrating from JavaEE to JakartaEE.

- update Quarkus bom from io.quarkus:quarkus-universe-bom to
io.quarkus.platform:quarkus-bom as it was deprecated since Quarkus 2.1
https://github.com/quarkusio/quarkus/wiki/Migration-Guide-2.1#quarkus-universe-bom-is-deprecated
(and it is required to use the Quarkus migration tool)
- launched quarkus `quarkus update --stream=3.0` as mentioned in
https://github.com/quarkusio/quarkus/wiki/Migration-Guide-3.0#automatic-update-tool
which dealt with a fair part of the migration. Then it remained few
little tasks.
- Updated quarkus.index-dependency for jaxb from
org.jboss.spec.javax.xml.bind:jboss-jaxb-api_2.3_spec to
`jakarta.xml.bind:jakarta.xml.bind-api`

fix #333

TO UPDATE THE version of compas.core when a release is available or use
the ${project.version} ?

Signed-off-by: Aurélien Pupier <apupier@redhat.com>